Remove koans directory from source control.
Patches were submitted against the koans directory rather than the
src directory. This lead to potential problems when we regenerate the
koans directory from scratch, leading to the very real possibility
that changes could be lost.

Please make all changes to the src directory and use "rake gen" (or
"rake regen") to generate the koans directory as needed.