Ignore RVM config if the user wants to use it
RVM allows the user to select a specific Ruby for a specific
project. It is common to put a .rvmrc config file in the root
of a project to select the Ruby automatically when the user
switches into that project.

The Ruby Koans should ignore this file if present and not mistake
it for a change to the Koans themselves. There are other ways
of choosing a Ruby to use, and use of RVM is not mandated by the
Koans so ignoring seems like the right solution.