Add a koan for mandatory keyword arguments
Previously, `about_keyword_arguments.rb` asked the reader
to reflect on the fact that keyword arguments must have default
values. However, this does not seem to be true anymore in Ruby.

This commit replaces that comment with two koans that guide the
reader through validating and using a method with mandatory
keyword arguments.